inlined implementation to header files and removed superfluous definitions
N.B. the aim is to make the disruptor also a header-only library that can be templated and more efficiently inlined/constexpr-optimised by the compiler
