+++ src/disruptor/include/disruptor/FixedSequenceGroup.hpp
@@ -1,9 +1,12 @@
 #pragma once
 
 #include <memory>
+#include <ostream>
 #include <vector>
 
+#include "Exceptions.hpp"
 #include "ISequence.hpp"
+#include "Util.hpp"
 
 namespace opencmw::disruptor {
 
@@ -19,34 +22,58 @@ public:
      *
      * \param sequences sequences the list of sequences to be tracked under this sequence group
      */
-    explicit FixedSequenceGroup(const std::vector<std::shared_ptr<ISequence>> &sequences);
+    explicit FixedSequenceGroup(const std::vector<std::shared_ptr<ISequence>> &sequences)
+        : m_sequences(sequences) {
+    }
 
     /**
      * Get the minimum sequence value for the group.
      */
-    std::int64_t value() const override;
+    std::int64_t value() const override {
+        return Util::getMinimumSequence(m_sequences);
+    }
 
     /**
      * Not supported.
      */
-    void setValue(std::int64_t value) override;
+    void setValue(std::int64_t /*value*/) override {
+        DISRUPTOR_THROW_NOT_SUPPORTED_EXCEPTION();
+    }
 
     /**
      * Not supported.
      */
-    bool compareAndSet(std::int64_t expectedValue, std::int64_t newValue) override;
+    bool compareAndSet(std::int64_t /*expectedValue*/, std::int64_t /*newValue*/) override {
+        DISRUPTOR_THROW_NOT_SUPPORTED_EXCEPTION();
+    }
 
     /**
      * Not supported.
      */
-    std::int64_t incrementAndGet() override;
+    std::int64_t incrementAndGet() override {
+        DISRUPTOR_THROW_NOT_SUPPORTED_EXCEPTION();
+    }
 
     /**
      * Not supported.
      */
-    std::int64_t addAndGet(std::int64_t increment) override;
+    std::int64_t addAndGet(std::int64_t /*increment*/) override {
+        DISRUPTOR_THROW_NOT_SUPPORTED_EXCEPTION();
+    }
 
-    void         writeDescriptionTo(std::ostream &stream) const override;
+    void writeDescriptionTo(std::ostream &stream) const override {
+        auto firstItem = true;
+        for (auto &&sequence : m_sequences) {
+            if (firstItem) {
+                firstItem = false;
+            } else {
+                stream << ", ";
+            }
+            stream << "{ ";
+            sequence->writeDescriptionTo(stream);
+            stream << " }";
+        }
+    }
 };
 
 } // namespace opencmw::disruptor
