+++ src/disruptor/include/disruptor/ProcessingSequenceBarrier.hpp
@@ -3,6 +3,9 @@
 #include <atomic>
 #include <vector>
 
+#include "Exceptions.hpp"
+#include "FixedSequenceGroup.hpp"
+#include "IHighestPublishedSequenceProvider.hpp"
 #include "ISequenceBarrier.hpp"
 #include "IWaitStrategy.hpp"
 #include "Sequence.hpp"
@@ -36,23 +39,61 @@ public:
     ProcessingSequenceBarrier(const std::shared_ptr<IHighestPublishedSequenceProvider> &sequenceProvider,
             const std::shared_ptr<IWaitStrategy>                                       &waitStrategy,
             const std::shared_ptr<Sequence>                                            &cursorSequence,
-            const std::vector<std::shared_ptr<ISequence>>                              &dependentSequences);
+            const std::vector<std::shared_ptr<ISequence>>                              &dependentSequences)
+        : m_waitStrategy(waitStrategy)
+        , m_dependentSequence(getDependentSequence(cursorSequence, dependentSequences))
+        , m_cursorSequence(cursorSequence)
+        , m_sequenceProvider(sequenceProvider)
+        , m_waitStrategyRef(*m_waitStrategy)
+        , m_dependentSequenceRef(*m_dependentSequence)
+        , m_cursorSequenceRef(*m_cursorSequence)
+        , m_sequenceProviderRef(*m_sequenceProvider) {
+        m_alerted = false;
+    }
 
-    std::int64_t waitFor(std::int64_t sequence) override;
+    std::int64_t waitFor(std::int64_t sequence) override {
+        checkAlert();
 
-    std::int64_t cursor() override;
+        auto availableSequence = m_waitStrategyRef.waitFor(sequence, *m_cursorSequence, *m_dependentSequence, *shared_from_this());
 
-    bool         isAlerted() override;
+        if (availableSequence < sequence) {
+            return availableSequence;
+        }
 
-    void         alert() override;
+        return m_sequenceProviderRef.getHighestPublishedSequence(sequence, availableSequence);
+    }
 
-    void         clearAlert() override;
+    std::int64_t cursor() override {
+        return m_dependentSequenceRef.value();
+    }
 
-    void         checkAlert() override;
+    bool isAlerted() override {
+        return m_alerted;
+    }
+
+    void alert() override {
+        m_alerted = true;
+        m_waitStrategyRef.signalAllWhenBlocking();
+    }
+
+    void clearAlert() override {
+        m_alerted = false;
+    }
+
+    void checkAlert() override {
+        if (m_alerted) {
+            DISRUPTOR_THROW_ALERT_EXCEPTION();
+        }
+    }
 
 private:
-    static std::shared_ptr<ISequence> getDependentSequence(const std::shared_ptr<Sequence> &cursorSequence,
-            const std::vector<std::shared_ptr<ISequence>>                                  &dependentSequences);
+    static std::shared_ptr<ISequence> getDependentSequence(const std::shared_ptr<Sequence> &cursorSequence, const std::vector<std::shared_ptr<ISequence>> &dependentSequences) {
+        if (dependentSequences.empty()) {
+            return cursorSequence;
+        }
+
+        return std::make_shared<FixedSequenceGroup>(dependentSequences);
+    }
 };
 
 } // namespace opencmw::disruptor
