+++ src/disruptor/include/disruptor/RoundRobinThreadAffinedTaskScheduler.hpp
@@ -7,7 +7,9 @@
 #include <vector>
 
 #include "BlockingQueue.hpp"
+#include "Exceptions.hpp"
 #include "ITaskScheduler.hpp"
+#include "ThreadHelper.hpp"
 
 namespace opencmw::disruptor {
 
@@ -18,18 +20,70 @@ class RoundRobinThreadAffinedTaskScheduler : public ITaskScheduler {
 private:
     BlockingQueue<std::packaged_task<void()>> m_tasks;
     std::atomic<bool>                         m_started{ false };
-    std::vector<std::thread>                  m_threads;
+    std::vector<std::jthread>                 m_threads;
 
 public:
-    void              start(std::int32_t numberOfThreads) override;
-    void              stop() override;
+    void start(std::size_t numberOfThreads) override {
+        if (m_started) {
+            return;
+        }
 
-    std::future<void> scheduleAndStart(std::packaged_task<void()> &&task) override;
+        m_started = true;
+
+        if (numberOfThreads < 1) {
+            DISRUPTOR_THROW_ARGUMENT_OUT_OF_RANGE_EXCEPTION(numberOfThreads);
+        }
+
+        createThreads(numberOfThreads);
+    }
+    void stop() override {
+        if (!m_started) {
+            return;
+        }
+
+        m_started = false;
+
+        for (auto &&thread : m_threads) {
+            if (thread.joinable()) {
+                thread.join();
+            }
+        }
+    }
+
+    std::future<void> scheduleAndStart(std::packaged_task<void()> &&task) override {
+        auto future = task.get_future();
+        m_tasks.push(std::move(task));
+
+        return future;
+    }
 
 private:
-    void createThreads(std::int32_t numberOfThreads);
-    void workingLoop(std::int32_t threadId);
-    void tryExecuteTask(std::packaged_task<void()> &task);
+    void createThreads(std::size_t numberOfThreads) {
+        for (auto i = 0U; i < numberOfThreads; ++i) {
+            m_threads.emplace_back([this, i]() { workingLoop(i); });
+        }
+    }
+
+    void workingLoop(std::size_t threadId) {
+        static const auto processorCount = std::thread::hardware_concurrency();
+
+        const auto        processorIndex = threadId % processorCount;
+
+        const auto        affinityMask   = ThreadHelper::AffinityMask(1ull << processorIndex);
+
+        ThreadHelper::setThreadAffinity(affinityMask);
+
+        while (m_started) {
+            std::packaged_task<void()> task;
+            while (m_tasks.timedWaitAndPop(task, std::chrono::milliseconds(100))) {
+                tryExecuteTask(task);
+            }
+        }
+    }
+
+    void tryExecuteTask(std::packaged_task<void()> &task) {
+        task();
+    }
 };
 
 } // namespace opencmw::disruptor
