+++ src/disruptor/include/disruptor/SleepingWaitStrategy.hpp
@@ -1,8 +1,12 @@
 #pragma once
 
 #include <cstdint>
+#include <ostream>
+#include <thread>
 
+#include "ISequenceBarrier.hpp"
 #include "IWaitStrategy.hpp"
+#include "Sequence.hpp"
 
 namespace opencmw::disruptor {
 
@@ -16,25 +20,48 @@ private:
     std::int32_t              m_retries        = 0;
 
 public:
-    explicit SleepingWaitStrategy(std::int32_t retries = m_defaultRetries);
+    explicit SleepingWaitStrategy(std::int32_t retries = m_defaultRetries)
+        : m_retries(retries) {
+    }
 
     /**
      * \see IWaitStrategy::waitFor()
      */
     std::int64_t waitFor(std::int64_t sequence,
-            Sequence                 &cursor,
-            ISequence                &dependentSequence,
-            ISequenceBarrier         &barrier) override;
+            Sequence & /*cursor*/,
+            ISequence        &dependentSequence,
+            ISequenceBarrier &barrier) override {
+        std::int64_t availableSequence;
+        auto         counter = m_retries;
+
+        while ((availableSequence = dependentSequence.value()) < sequence) {
+            counter = applyWaitMethod(barrier, counter);
+        }
+
+        return availableSequence;
+    }
 
     /**
      * \see IWaitStrategy::signalAllWhenBlocking()
      */
-    void signalAllWhenBlocking() override;
-
-    void writeDescriptionTo(std::ostream &stream) const override;
+    void signalAllWhenBlocking() override {}
+    void writeDescriptionTo(std::ostream &stream) const override { stream << "SleepingWaitStrategy"; }
 
 private:
-    static std::int32_t applyWaitMethod(ISequenceBarrier &barrier, std::int32_t counter);
+    static std::int32_t applyWaitMethod(ISequenceBarrier &barrier, std::int32_t counter) {
+        barrier.checkAlert();
+
+        if (counter > 100) {
+            --counter;
+        } else if (counter > 0) {
+            --counter;
+            std::this_thread::yield();
+        } else {
+            std::this_thread::sleep_for(std::chrono::milliseconds(0));
+        }
+
+        return counter;
+    }
 };
 
 } // namespace opencmw::disruptor
