+++ src/disruptor/include/disruptor/SpinWaitWaitStrategy.hpp
@@ -1,6 +1,11 @@
 #pragma once
 
+#include <ostream>
+
+#include "ISequenceBarrier.hpp"
 #include "IWaitStrategy.hpp"
+#include "Sequence.hpp"
+#include "SpinWait.hpp"
 
 namespace opencmw::disruptor {
 
@@ -14,16 +19,25 @@ public:
      * \see IWaitStrategy.waitFor()
      */
     std::int64_t waitFor(std::int64_t sequence,
-            Sequence                 &cursor,
-            ISequence                &dependentSequence,
-            ISequenceBarrier         &barrier) override;
+            Sequence & /*cursor*/,
+            ISequence        &dependentSequence,
+            ISequenceBarrier &barrier) override {
+        std::int64_t availableSequence;
+
+        SpinWait     spinWait;
+        while ((availableSequence = dependentSequence.value()) < sequence) {
+            barrier.checkAlert();
+            spinWait.spinOnce();
+        }
+
+        return availableSequence;
+    }
 
     /**
      * \see IWaitStrategy.signalAllWhenBlocking()
      */
-    void signalAllWhenBlocking() override;
-
-    void writeDescriptionTo(std::ostream &stream) const override;
+    void signalAllWhenBlocking() override {}
+    void writeDescriptionTo(std::ostream &stream) const override { stream << "SpinWaitWaitStrategy"; }
 };
 
 } // namespace opencmw::disruptor
