+++ src/disruptor/include/disruptor/ThreadHelper_Windows.cpp
@@ -1,7 +1,4 @@
-#include "stdafx.hpp"
 #include "ThreadHelper.hpp"
-
-#include "BuildConfig.hpp"
 
 #ifdef DISRUPTOR_OS_FAMILY_WINDOWS
