+++ src/disruptor/include/disruptor/Util.hpp
@@ -21,14 +21,24 @@ namespace Util {
  * \param x Value to round up
  * \returns The next power of 2 from x inclusive
  */
-std::int32_t ceilingNextPowerOfTwo(std::int32_t x);
+inline std::int32_t ceilingNextPowerOfTwo(std::int32_t x) {
+    std::int32_t result = 2;
+
+    while (result < x) {
+        result <<= 1;
+    }
+
+    return result;
+}
 
 /**
  * Test whether a given integer is a power of 2
  *
  * \param x
  */
-bool isPowerOf2(std::int32_t x);
+inline bool isPowerOf2(std::int32_t x) {
+    return x > 0 && (x & (x - 1)) == 0;
+}
 
 /**
  * Calculate the log base 2 of the supplied integer, essentially reports the location of the highest bit.
@@ -36,7 +46,13 @@ bool isPowerOf2(std::int32_t x);
  * \param i Value to calculate log2 for.
  * \returns The log2 value
  */
-std::int32_t log2(std::int32_t i);
+inline std::int32_t log2(std::int32_t i) {
+    std::int32_t r = 0;
+    while ((i >>= 1) != 0) {
+        ++r;
+    }
+    return r;
+}
 
 /**
  * Get the minimum sequence from an array of Sequences.
@@ -45,7 +61,14 @@ std::int32_t log2(std::int32_t i);
  * \param minimum an initial default minimum.  If the array is empty this value will returned.
  * \returns the minimum sequence found or lon.MaxValue if the array is empty.
  */
-std::int64_t getMinimumSequence(const std::vector<std::shared_ptr<ISequence>> &sequences, std::int64_t minimum = std::numeric_limits<std::int64_t>::max());
+inline std::int64_t getMinimumSequence(const std::vector<std::shared_ptr<ISequence>> &sequences, std::int64_t minimum = std::numeric_limits<std::int64_t>::max()) {
+    if (sequences.empty()) {
+        return minimum;
+    } else {
+        return std::min(minimum,
+                std::ranges::min(sequences, std::less{}, [](auto &&sequence) { return sequence->value(); })->value());
+    }
+}
 
 /**
  * Get an array of Sequences for the passed IEventProcessors
@@ -53,7 +76,12 @@ std::int64_t getMinimumSequence(const std::vector<std::shared_ptr<ISequence>> &s
  * \param processors processors for which to get the sequences
  * \returns the array of\returns <see cref="Sequence"/>\returns s
  */
-std::vector<std::shared_ptr<ISequence>> getSequencesFor(const std::vector<std::shared_ptr<IEventProcessor>> &processors);
+inline std::vector<std::shared_ptr<ISequence>> getSequencesFor(const std::vector<std::shared_ptr<IEventProcessor>> &processors) {
+    // Ah C++20 ranges, no conversions to vector yet
+    std::vector<std::shared_ptr<ISequence>> sequences(processors.size());
+    std::ranges::transform(processors, sequences.begin(), [](auto &&processor) { return processor->sequence(); });
+    return sequences;
+}
 
 } // namespace Util
 } // namespace opencmw::disruptor
