+++ src/disruptor/include/disruptor/YieldingWaitStrategy.hpp
@@ -1,6 +1,11 @@
 #pragma once
 
+#include <ostream>
+#include <thread>
+
+#include "ISequenceBarrier.hpp"
 #include "IWaitStrategy.hpp"
+#include "Sequence.hpp"
 
 namespace opencmw::disruptor {
 
@@ -23,19 +28,37 @@ public:
      * \param barrier barrier the IEventProcessor is waiting on.
      * \returns the sequence that is available which may be greater than the requested sequence.
      */
-    std::int64_t waitFor(std::int64_t sequence,
-            Sequence                 &cursor,
-            ISequence                &dependentSequence,
-            ISequenceBarrier         &barrier) override;
+    std::int64_t waitFor(std::int64_t sequence, Sequence & /*cursor*/, ISequence &dependentSequence, ISequenceBarrier &barrier) override {
+        std::int64_t availableSequence;
+        auto         counter = m_spinTries;
+
+        while ((availableSequence = dependentSequence.value()) < sequence) {
+            counter = applyWaitMethod(barrier, counter);
+        }
+
+        return availableSequence;
+    }
 
     /**
      * Signal those IEventProcessor waiting that the cursor has advanced.
      */
-    void                signalAllWhenBlocking() override;
+    void                signalAllWhenBlocking() override {}
 
-    static std::int32_t applyWaitMethod(ISequenceBarrier &barrier, std::int32_t counter);
+    static std::int32_t applyWaitMethod(ISequenceBarrier &barrier, std::int32_t counter) {
+        barrier.checkAlert();
 
-    void                writeDescriptionTo(std::ostream &stream) const override;
+        if (counter == 0) {
+            std::this_thread::yield();
+        } else {
+            --counter;
+        }
+
+        return counter;
+    }
+
+    void writeDescriptionTo(std::ostream &stream) const override {
+        stream << "YieldingWaitStrategy";
+    }
 };
 
 } // namespace opencmw::disruptor
