--- src/disruptor/include/disruptor/stdafx.cpp
@@ -1,8 +1,0 @@
-// stdafx.cpp : source file that includes just the standard includes
-// disruptorLib.pch will be the pre-compiled header
-// stdafx.obj will contain the pre-compiled type information
-
-#include "stdafx.hpp"
-
-// TODO: reference any additional headers you need in STDAFX.H
-// and not in this file
