--- src/disruptor/include/disruptor/targetver.hpp
@@ -1,2 +1,0 @@
-#pragma once
-
